requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
Pillow==10.1.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
django-storages==1.14.2
django-filter==23.5
djangorestframework-simplejwt==5.3.1
channels==4.0.0
black==24.10.0
isort==5.13.2
whitenoise==6.7.0
uvicorn==0.31.0
gunicorn==23.0.0
pre_commit==4.0.1
